consumer/requirements.txt
kafka-python==2.3.0
psycopg2-binary==2.9.11